[fix] In legacy mailbox protocol, look at response size.
Some callers expect libhoth_receive_response() to return the actual size
of the response in `*actual_size`, which is what the new FIFO protocol does
(since there is lower-level framing). The mailbox implementation should
be consistent with this. To do this, we need to look at the response
header to determine the size. We set `*actual_size` to that value, and
since we now know the response size there's no point reading any further
than we to.

In addition to making the API consistent, this also speeds up "htool spi
update" by as much as 50%.