htool console: Don't leave O_NONBLOCK set on stdin
"htool console" sets O_NONBLOCK on stdin to allow it to poll for
incoming keystrokes. However, on some TTYs, O_NONBLOCK on stdin will
also apply to stdout, which will cause it to drop data when the TTY's
buffer is full.

Also, improve error handling if the stdout write fails.