[protocol] Add remaining unit test for auth record
We are seeking to fill in missing code coverage for the auth record for libhoth.
This will coverage some negative test cases that we should expect to see fail give
known specific values. This includes that failing case of a build record. In addition,
to a mismatch key id with the auth record

Signed-off-by: Ellis Sarza-Nguyen <sarzanguyen@google.com>